backend/py/requirements.txt
# FastAPI Backend Requirements for Churn Analysis
# Python 3.11+ compatible

# Core FastAPI dependencies
fastapi==0.104.*
uvicorn[standard]==0.24.*

# Database connectivity
asyncpg==0.29.*
psycopg[binary]==3.1.*

# Data processing
pandas==2.1.*
numpy==1.25.*

# Scientific computing (for ML tasks)
scikit-learn==1.3.*
joblib==1.3.*

# Environment and configuration
python-dotenv==1.0.*

# AI and OpenAI integration
openai==1.12.*

# Development and testing
pytest==7.4.*
pytest-asyncio==0.21.*

# Optional: Enhanced logging and monitoring
structlog==23.2.*

# Optional: API documentation enhancements
pydantic==2.5.*